fix: Correct relative import paths in moved utility files
Fixes 3 critical bugs found in code review (Copilot):
- concurrency-limiter.ts: Fix import depth for alexandria-cover-service
- kv-results-handler.ts: Fix import depth for response-builder
- book-metadata.ts: Fix import depth for types/env

These files were moved one level deeper but imports weren't updated.
All 199 smoke tests passing.

Resolves review comments from PR #235.
